[#363] Fixed compiler error caused by NUMOFLOW not reported
NUMOFLOWs are quietly ignored during s2n in compile time, and this
wasn't being detected by the numeric optimizations. As a result,
the numeric optimization would continue without the numeric
representation of a given string value being filled out. In many cases,
this meant that a value of '0' would be used, resulting in incorrect
behavior for boolean expressions.