[AARCH64] Temporarily disable gcc LTO (LinkTimeOptimization) when building YottaDB on 64-bit ARM as it causes SIG-11
We had a few subtests (m_commands/readwt and mpt_extra/gbl) that failed with SIG-11 on 64-bit ARM.
Below is the failure in the mpt_extra/gbl subtest.

sr_port/gvzwr_var.c
--------------------
    220      if (NEVER == gv_cur_region->null_subs || seen_null || gv_cur_region->std_null_coll)

Machine instructions
---------------------
    |0xffffa4bd91a4 <gvzwr_var+628>  ldr    w0, [x29, #164]                    |
    |0xffffa4bd91a8 <gvzwr_var+632>  cbnz   w0, 0xffffa4bd924c <gvzwr_var+796> |
    |0xffffa4bd91ac <gvzwr_var+636>  ldr    x0, [x27, #1208]                   |
 -->|0xffffa4bd91b0 <gvzwr_var+640>  ldrsb  w1, [x0, #58]                      |
    |0xffffa4bd91b4 <gvzwr_var+644>  cmp    w1, #0x0                           |

The SIG-11 was at the ldrsb instruction because $x0 register contains 0. #58 is the offset of
"null_subs" in the "gd_region" structure.  In the corresponding C code, we are trying to dereference
gv_cur_region->null_subs at line 220 above.

In the core, the global variable gv_cur_region is a valid pointer. But somehow the pro build
is accessing a wrong address to get at the pointer value of the global variable "gv_cur_region".
I suspect it is a C compiler optimization issue on 64-bit ARM with using LTO.  When LTO is disabled
the same test case reliably passes whereas it fails reliably with LTO.

Note that disabling LTO increased the libyottadb.so size from 4.1Mb to 4.7Mb so it is desirable to
re-enable LTO as soon as possible (i.e. when gcc LTO gets fixed). This will be periodically checked
and re-enabled when things start working again.

When keeping LTO enabled but using ld (instead of gold), the build on 64-bit ARM showed an error in
save_gv_cur_region variable being multiply defined and true enough it had two STATICDEF definitions.
So that is now fixed by renaming one of the two definitions (one in gvn2gds.c) with a gvn2gds_ prefix
to avoid the duplicate definition. Even though LTO is not enabled currently, this is a good change to
do independent of LTO and/or whether ld or gold is used as the linker.