[#588] MUPIP LOAD needs to warn if zwrite/go format extract file is missing 2-line header
Added a way to check if header is present or not in extracted file in
mupip_cbtgbl.c:get_load_format. If no header, read the file
from the start of the file and load the file to yottadb. This is done by
resetting the file read position of the extracted file to the start of
the file.