[#497] ydbinstall should not require --force-install to install a distribution built on the same system
Rearranged the ydb-install to ensure that platform and architecture
information is checked before the ydb-force install check and skips it
if the builds platform and architecture matches current environment
platform and architecture changes.

Currently, the ydbinstall script checks to see if the build system and
the currently downloaded tar files build information matches and
continue if it does.

Todo: Need to add logic to get system information from the
etc/os-release and check it against the build os information from the
remote github repo.