[#588] WIP Different approach for ydb 588
In this approach the header files are not read twice but rather managed
through pointer manipulation. Since, named pipes could not be read
twice.

mupip_cvtgbl.c -> get_load_format returns 3 pointers containing the
first 3 lines if it is zwr and 2 lines if it is go file. This is passed
onto the go_load.c file and is added to the database. In order to
accomodate the first two/three lines (based on the format).

The get_load_format checks if the first line contains zwr/go format or
if the command has explicitly specified a format eg: mupip load
-format=go/zwr. If not then, it checks that the first two lines follow
the valid regex of the two formats and specifies the format. If it is of
valid go/zwr format the headless flag is enabled and format is
specified, or else it is invalid format.

The go_load.c file checks for headless flag and prevents load from the file
for the first two lines, since they will be added to the database from
the pointers.

What is remaining?
Testing the code with mu_load_err file. All other tests work fine.