[V63011] Fixes to V63011 imported changes
Several of the changes within undid changes that had been done by YottaDB so those changes
are restored in this commit.

Most of the rest of the changes are changes to GDE when it tries to create a global directory
in a subdirectory where the path name plus "mumps.gld" creates a path longer than 255 bytes
which YottaDB can not currently deal with. This now pushes out a new error (INVGBLDIR2) that
is a variant of the INVGBLDIR message that does not "continue with mumps.gld".