[V63011] Adding support for SIGHUP in alternate signal handling
This support was added to normal signal handling in V63011 and is controlled by device parameter. We
also need to support it in the alternate signaling handling scheme though there, since SIGHUP is mostly
treated as a ^C, the alternate signal handling scheme just treats SIGHUP as if it were a SIGTERM, which
is how ^C is treated in SimpleAPI calls.