[#850] Shrink Rocky Linux Docker yottadb/yottadb-rocky image
Shrink down from 571MB -> 364MB.

Tricks used to accomplish this:

1. Make a YottaDB Release build, just like the Ubuntu Image. A
RelWithDbgInfo build is almost twice the size (65MB -> 130MB).
2. Use Docker Buildkit to mount a layer from one image to another rather
than copying it.
3. Only install `libicu` instead of `libicu-devel`
4. Don't install `glibc-langpack-en`. Rely on C UTF-8 locale.
5. Install `nano` instead of `vim`.
6. Temporarily install wget and then remove it.
7. Clean yum stuff after we are done

Unfortunately, I couldn't get it down to the size of the Ubuntu image.
This is mainly because the Ubuntu Image is 72.8MB to start with; but
Rocky Linux is 205MB.